12_reflect/walk/walk.go
package walk

import (
	"reflect"
)

func walk(x interface{}, fn func(string)) {
	// fn("IIII")

	// val := reflect.ValueOf(x)
	// field := val.Field(0)
	// fn(field.String())

	val := getValue(x)
	// numOfValues := 0
	// var getField func(i int) reflect.Value
	walkValue := func(value reflect.Value) {
		walk(value.Interface(), fn)
	}

	switch val.Kind() {
	case reflect.Slice, reflect.Array:
		for i := 0; i < val.Len(); i++ {
			walkValue(val.Index(i))
		}
	case reflect.Struct:
		for i := 0; i < val.NumField(); i++ {
			walkValue(val.Field(i))
		}
	case reflect.Map:
		for _, key := range val.MapKeys() {
			walkValue(val.MapIndex(key))
		}
	case reflect.String:
		fn(val.String())
	}
}

func getValue(x interface{}) reflect.Value {
	val := reflect.ValueOf(x)
	if val.Kind() == reflect.Pointer {
		val = val.Elem()
	}
	return val
}
